Add ability to listen to all messages (wildcard)
This commit implements a special method value '*' that, when used in a
call to the 'on' method, makes the listener function receive all remote
prodecure calls. Regular listeners still receive their matching call,
after the call has been passed to the wildcard listener. If a regular
listener matches a call, that listener's response is used. If not, the
response from the wildcard listener is used instead.

This commit also includes tests and README.md change covering the new
functionality.